Refactor taskAction resource to take in a Task definition uuid in addition to name
