+++ omod-2.4/src/main/java/org/openmrs/module/webservices/helper/TaskServiceWrapper2_4.java
@@ -1,0 +1,21 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ *
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs.module.webservices.helper;
+
+import org.openmrs.api.context.Context;
+import org.openmrs.scheduler.TaskDefinition;
+
+public class TaskServiceWrapper2_4 extends TaskServiceWrapper {
+	
+	public TaskDefinition getTaskByUuid(String uuid) {
+		return Context.getSchedulerService().getTaskByUuid(uuid);
+	}
+	
+}
